Move Suunto DB imports into parse-db.c
